train.py: add user config

+++ train.py
@@ -8,6 +8,7 @@ import math
 import random
 import shutil
 import pathlib
+import tempfile
 import mido
 import tqdm
 import torch
@@ -16,6 +17,7 @@ import torch.nn.functional as F
 import matplotlib.pyplot as plt
 from torch.utils.data import Dataset, DataLoader
 from torch.nn import DataParallel
+from functools import partial
 
 import warnings
 import os
@@ -246,14 +248,9 @@ def train(model: MidiNet, dataset: MidiDataset, optimizer: optim.SGD, train_batc
             loss.backward()  # 反向传播
             optimizer.step()  # 更新模型参数
 
-            # 计算训练准确率
-            accuracy = (torch.argmax(outputs) == labels).sum().item()
             train_loss_sum += loss.item()  # 累积训练损失
-            train_accuracy_sum += accuracy  # 累积训练准确率
-
-            # 更新进度条
-            progress_bar.update()
-            progress_bar.set_postfix(loss=loss.item(), accuracy=f"{accuracy / labels.size(0):.2f}")
+            train_accuracy_sum += (torch.argmax(outputs) == labels).sum().item()  # 累积训练准确率
+            progress_bar.update()  # 更新进度条
 
             # 验证阶段
             if step % val_per_step == 0:
@@ -350,32 +347,49 @@ def main():
     """
     训练 MIDI 模型并绘制训练过程中的损失曲线。
     """
+    # 用户定义的配置字典
+    config = {
+        "pretrained_ckpt": "/kaggle/input/tkaimidi/pytorch/default/1/ckpt",  # 预训练模型的检查点路径，可以为空
+        "local_ckpt": "ckpt",  # 在本地保存的检查点的路径，训练结束后会保存到这里
+        "external_datasets_path": ["/kaggle/input/music-midi"],  # 外部数据集的路径列表
+        "lr": 0.003548133892335713,  # 学习率
+        "weight_decay": 1e-3,  # 权重衰减系数
+        "seq_size": 512,  # 输入序列的大小 - 1
+        "val_steps": 1,  # 进行多少次验证步骤（因为 Dataset 进行了数据增强，一个 Epoch 训练数据变得很多，Kaggle GPU 12个小时跑不完，所以用验证步骤代替）
+        "train_batch_size": 1,  # 训练时的批量大小
+        "val_batch_size": 4,  # 验证时的批量大小，越大验证结果越准确，但是资源使用倍数增加，验证时间也增加（但没有资源使用增加得多）
+        "train_length": 0.8,  # 训练集占数据集的比例，用来保证用来验证的数据不被训练
+        "val_per_step": 4096,  # 每多少个训练步骤进行一次验证
+        "steps_to_val": 256  # 抽样验证，每一次验证使用多少个批次，越大验证结果越准确，但是验证时间倍数增加，资源使用不增加
+    }
+
     # 定义路径
-    pretrained_ckpt = pathlib.Path("/kaggle/input/tkaimidi/pytorch/default/1/ckpt")
-    local_ckpt = pathlib.Path("ckpt")
-    local_dataset_path = pathlib.Path("/kaggle/temp/data")
-    external_datasets_path = [pathlib.Path(path) for path in ["/kaggle/input/music-midi"]]
+    local_ckpt = pathlib.Path(config["local_ckpt"])
+    local_dataset_path = pathlib.Path(tempfile.gettempdir())
+    external_datasets_path = [pathlib.Path(path) for path in config["external_datasets_path"]]
 
     # 如果预训练检查点存在，则复制到本地检查点路径
-    if pretrained_ckpt.exists():
+    if config["pretrained_ckpt"] and pathlib.Path(config["pretrained_ckpt"]).exists():
         if local_ckpt.exists():
             shutil.rmtree(local_ckpt)  # 删除现有的本地检查点
-        shutil.copytree(pretrained_ckpt, local_ckpt, dirs_exist_ok=True)  # 复制预训练检查点到本地
+        shutil.copytree(config["pretrained_ckpt"], local_ckpt, dirs_exist_ok=True)  # 复制预训练检查点到本地
 
     # 复制外部数据集到本地
     for path in external_datasets_path:
         shutil.copytree(path, local_dataset_path / path.name, dirs_exist_ok=True)
 
-    dataset = MidiDataset(local_dataset_path, seq_size=512)  # 加载训练数据集
+    dataset = MidiDataset(local_dataset_path, seq_size=config["seq_size"])  # 加载训练数据集
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")  # 获取设备
     model = MidiNet()  # 初始化模型
 
     # 尝试加载检查点
+    create_optimizer = partial(optim.SGD, model.parameters(), lr=config["lr"], momentum=0.9, weight_decay=config["weight_decay"])
+
     try:
         model_state, optimizer_state, old_train_loss, old_val_loss, old_train_accuracy, old_val_accuracy, dataset_length, train_start, last_batch, generator_state = load_checkpoint(local_ckpt, optimizer=True)
         model.load_state_dict(model_state)  # 加载模型状态
         model = model.to(device)  # 转移到指定设备并编译模型
-        optimizer = optim.SGD(model.parameters(), lr=0.003548133892335713, momentum=0.9, weight_decay=1e-3)  # 初始化优化器
+        optimizer = create_optimizer()  # 初始化优化器
         optimizer.load_state_dict(optimizer_state)
 
         if dataset_length != len(dataset):
@@ -385,14 +399,14 @@ def main():
             last_batch = 0
     except Exception as e:
         print(f"加载检查点时发生错误: {e}", file=sys.stderr)
-        optimizer = optim.SGD(model.parameters(), lr=0.003548133892335713, momentum=0.9, weight_decay=1e-3)  # 初始化优化器
+        optimizer = create_optimizer()  # 初始化优化器
         old_train_loss, old_val_loss, old_train_accuracy, old_val_accuracy = [], [], [], []
         train_start = random.randint(0, len(dataset) - 1)
         generator_state = ...
         last_batch = 0
 
     # 开始训练模型
-    train_loss, val_loss, train_accuracy, val_accuracy, last_batch, generator_state = train(model, dataset, optimizer, val_steps=1, train_batch_size=1, val_batch_size=4, train_start=train_start, generator_state=generator_state, last_batch=last_batch, device=device)
+    train_loss, val_loss, train_accuracy, val_accuracy, last_batch, generator_state = train(model, dataset, optimizer, val_steps=config["val_steps"], train_batch_size=config["train_batch_size"], val_batch_size=config["val_batch_size"], train_length=config["train_length"], val_per_step=config["val_per_step"], steps_to_val=config["steps_to_val"], train_start=train_start, generator_state=generator_state, last_batch=last_batch, device=device)
 
     # 合并旧的损失记录和新的记录
     train_loss = old_train_loss + train_loss
